requirements.txt
requests
bs4
fake_headers
fake_useragent
python-dotenv
